+++ src/main/kotlin/com/whatever/caro/card/internal/deck/dto/create/CreateDeckRequest.kt
@@ -1,4 +1,4 @@
-package com.whatever.caro.card.internal.deck.dto
+package com.whatever.caro.card.internal.deck.dto.create
 
 import io.swagger.v3.oas.annotations.media.Schema
 import jakarta.validation.constraints.NotBlank
